cleanup: post-refactoring code cleanup
After refactoring to "one pipelinerun per component", clean up
remaining legacy code and improve overall code organization:

- Remove pkg directory and relocate required code to internal/pkg
- Standardize code style, including package and variable naming
- Add license headers to source files
