+++ internal/controller/pipelinerun_controller_test.go
@@ -1,18 +1,16 @@
-/*
-Copyright 2024.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
+// Copyright 2024 Red Hat, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
 
 // What to test:
 // - check countRunningPipelineRuns
@@ -27,8 +25,8 @@ import (
 	. "github.com/onsi/gomega"
 	tektonv1 "github.com/tektoncd/pipeline/pkg/apis/pipeline/v1"
 
-	local_tekton "github.com/konflux-ci/mintmaker/internal/pkg/tekton"
-	. "github.com/konflux-ci/mintmaker/pkg/common"
+	. "github.com/konflux-ci/mintmaker/internal/pkg/constant"
+	tekton "github.com/konflux-ci/mintmaker/internal/pkg/tekton"
 )
 
 var _ = Describe("PipelineRun Controller", FlakeAttempts(3), func() {
@@ -42,7 +40,7 @@ var _ = Describe("PipelineRun Controller", FlakeAttempts(3), func() {
 
 			for i := range 3 {
 				pplrName := "pplnr" + strconv.Itoa(i)
-				pipelineRunBuilder := local_tekton.NewPipelineRunBuilder(pplrName, MintMakerNamespaceName)
+				pipelineRunBuilder := tekton.NewPipelineRunBuilder(pplrName, MintMakerNamespaceName)
 				pipelinerun, err := pipelineRunBuilder.Build()
 				Expect(err).NotTo(HaveOccurred())
 				Expect(k8sClient.Create(ctx, pipelinerun)).Should(Succeed())
@@ -86,13 +84,13 @@ var _ = Describe("PipelineRun Controller", FlakeAttempts(3), func() {
 			pplrName := "pplnr1"
 			labels := make(map[string]string)
 			labels[MintMakerAppstudioLabel] = "github"
-			pipelineRunBuilder := local_tekton.NewPipelineRunBuilder(pplrName, MintMakerNamespaceName)
+			pipelineRunBuilder := tekton.NewPipelineRunBuilder(pplrName, MintMakerNamespaceName)
 			pipelinerun, err := pipelineRunBuilder.WithLabels(labels).Build()
 			Expect(err).NotTo(HaveOccurred())
 			Expect(k8sClient.Create(ctx, pipelinerun)).Should(Succeed())
 
 			pplrName = "pplnr2"
-			pipelineRunBuilder = local_tekton.NewPipelineRunBuilder(pplrName, MintMakerNamespaceName)
+			pipelineRunBuilder = tekton.NewPipelineRunBuilder(pplrName, MintMakerNamespaceName)
 			pipelinerun, err = pipelineRunBuilder.Build()
 			Expect(err).NotTo(HaveOccurred())
 			Expect(k8sClient.Create(ctx, pipelinerun)).Should(Succeed())
