--- internal/pkg/common/constants.go
@@ -1,6 +1,0 @@
-package common
-
-const (
-	RenovateImageEnvName    = "RENOVATE_IMAGE"
-	DefaultRenovateImageURL = "quay.io/konflux-ci/mintmaker-renovate-image:latest"
-)
