+++ internal/pkg/utils/random.go
@@ -12,18 +12,17 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package main
+package utils
 
 import (
-	"flag"
-
-	osv_generator "github.com/konflux-ci/mintmaker/tools/osv-generator"
+	"crypto/rand"
+	"encoding/hex"
 )
 
-// A demo which parses RPM CVE data created in the last 24 hours into OSV database format
-func main() {
-	filename := flag.String("filename", "redhat.nedb", "Output filename for OSV database")
-	flag.Parse()
-
-	osv_generator.GenerateOSV(*filename)
+func RandomString(length int) string {
+	bytes := make([]byte, length/2+1)
+	if _, err := rand.Read(bytes); err != nil {
+		panic("Failed to read from random generator")
+	}
+	return hex.EncodeToString(bytes)[0:length]
 }
