--- pkg/common/random.go
@@ -1,14 +1,0 @@
-package common
-
-import (
-	"crypto/rand"
-	"encoding/hex"
-)
-
-func RandomString(length int) string {
-	bytes := make([]byte, length/2+1)
-	if _, err := rand.Read(bytes); err != nil {
-		panic("Failed to read from random generator")
-	}
-	return hex.EncodeToString(bytes)[0:length]
-}
