--- pkg/git/githubapp/app.go
@@ -1,10 +1,0 @@
-package githubapp
-
-import (
-	"context"
-)
-
-// ConfigReader is an interface for reading GitHub App configuration.
-type ConfigReader interface {
-	GetConfig(ctx context.Context) (githubAppIdStr string, appPrivateKeyPem []byte, err error)
-}
