--- pkg/git/gitproviderfactory/gitproviderfactory.go
@@ -1,139 +1,0 @@
-/*
-Copyright 2024 Red Hat, Inc.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package gitproviderfactory
-
-import (
-	"fmt"
-	"strconv"
-
-	. "github.com/konflux-ci/mintmaker/pkg/common"
-	"github.com/konflux-ci/mintmaker/pkg/git/github"
-	"github.com/konflux-ci/mintmaker/pkg/git/gitlab"
-	"github.com/konflux-ci/mintmaker/pkg/git/gitprovider"
-	"github.com/konflux-ci/mintmaker/pkg/mmerrors"
-)
-
-var CreateGitClient func(gitClientConfig GitClientConfig) (gitprovider.GitProviderClient, error) = createGitClient
-
-type GitClientConfig struct {
-	// PacSecretData are the content of Pipelines as Code secret
-	PacSecretData map[string][]byte
-	// GitProvider is type of the git provider to construct client for.
-	// Cannot be obtained from repo repository URL in case of self-hosted solution.
-	GitProvider string
-	// RepoUrl is the target git repository URL.
-	// Used to check that the requirements to access the repository are met,
-	// for example, check that the application is installed into given git repository.
-	// Ignored for some client configurations, e.g. clients created directly via a token.
-	RepoUrl string
-	// IsAppInstallationExpected shows whether to expect application installation into the target repository URL.
-	// Ignored for clients created directly via a token.
-	// Only for simple builds must be set to false.
-	IsAppInstallationExpected bool
-}
-
-// createGitClient creates new git provider client for the requested config
-func createGitClient(gitClientConfig GitClientConfig) (gitprovider.GitProviderClient, error) {
-	gitProvider := gitClientConfig.GitProvider
-	secretData := gitClientConfig.PacSecretData
-	username, usernameExists := secretData["username"]
-	password, passwordExists := secretData["password"]
-	_, sshKeyExists := secretData["ssh-privatekey"]
-
-	isAppUsed := IsPaCApplicationConfigured(gitProvider, secretData)
-
-	switch gitProvider {
-	case "github":
-		if !isAppUsed {
-			if passwordExists {
-				if usernameExists {
-					return github.NewGithubClientWithBasicAuth(string(username), string(password)), nil
-				}
-				return github.NewGithubClient(string(password)), nil
-			}
-			if sshKeyExists {
-				return nil, mmerrors.NewMintMakerError(mmerrors.EGitHubSecretTypeNotSupported,
-					fmt.Errorf("failed to create git client: GitHub ssh key authentication not yet supported"))
-			}
-			return nil, mmerrors.NewMintMakerError(mmerrors.EGitHubSecretTypeNotSupported,
-				fmt.Errorf("failed to create git client:  unsupported secret data. Expected username/password or token"))
-		}
-
-		githubAppIdStr := string(secretData[PipelinesAsCodeGithubAppIdKey])
-		githubAppId, err := strconv.ParseInt(githubAppIdStr, 10, 64)
-		if err != nil {
-			return nil, mmerrors.NewMintMakerError(mmerrors.EGitHubAppMalformedId,
-				fmt.Errorf("failed to create git client: failed to convert %s to int: %w", githubAppIdStr, err))
-		}
-
-		privateKey := secretData[PipelinesAsCodeGithubPrivateKey]
-
-		if gitClientConfig.IsAppInstallationExpected {
-			// It's required that the configured Pipelines as Code application is installed into user's account
-			// and enabled for the given repository.
-
-			githubClient, err := github.NewGithubClientByApp(githubAppId, privateKey, gitClientConfig.RepoUrl)
-			if err != nil {
-				return nil, err
-			}
-
-			// Check if the application is installed into target repository
-			appInstalled, err := githubClient.IsAppInstalledIntoRepository(gitClientConfig.RepoUrl)
-			if err != nil {
-				return nil, err
-			}
-			if !appInstalled {
-				return nil, mmerrors.NewMintMakerError(mmerrors.EGitHubAppNotInstalled,
-					fmt.Errorf("failed to create git client: GitHub Application is not installed into the repository"))
-			}
-			return githubClient, nil
-		} else {
-			// For simple builds we need to query repositories where configured Pipelines as Code application is not installed.
-			githubClient, err := github.NewGithubClientForSimpleBuildByApp(githubAppId, privateKey)
-			if err != nil {
-				return nil, err
-			}
-			return githubClient, nil
-		}
-
-	case "gitlab":
-		if isAppUsed {
-			return nil, fmt.Errorf("GitLab does not have applications")
-		}
-		baseUrl, err := gitlab.GetBaseUrl(gitClientConfig.RepoUrl)
-		if err != nil {
-			return nil, err
-		}
-		if usernameExists && passwordExists {
-			return gitlab.NewGitlabClientWithBasicAuth(string(username), string(password), baseUrl)
-		}
-		if !usernameExists && passwordExists {
-			return gitlab.NewGitlabClient(string(password), baseUrl)
-		}
-		if sshKeyExists {
-			return nil, mmerrors.NewMintMakerError(mmerrors.EGitLabSecretTypeNotSupported,
-				fmt.Errorf("failed to create git client: GitLab ssh key authentication not yet supported"))
-		}
-		return nil, mmerrors.NewMintMakerError(mmerrors.EGitLabSecretTypeNotSupported,
-			fmt.Errorf("failed to create git client: unsupported secret data. Expected username/password or token"))
-
-	case "bitbucket":
-		return nil, mmerrors.NewMintMakerError(mmerrors.EUnknownGitProvider, fmt.Errorf("git provider %s is not supported", gitProvider))
-	default:
-		return nil, mmerrors.NewMintMakerError(mmerrors.EUnknownGitProvider, fmt.Errorf("git provider %s is not supported", gitProvider))
-	}
-}
